sensor_plot_pc/requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.58.0
importlib-resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.17.0
zipp==3.21.0
